Refactor argument parsing and validation
This moves the responsibility of argument parsing to the toplevel
dispatch function (in combination with the request spec).  Appropriate
behavior in the face of bad arguments is left to the individual
handlers.

The primary change introduced is the introduction of support for
non-trailing variadic parameters.  Currently, they are not handled
correctly in the presence of subsequent optional arguments, but that
support might be added later.