modernize async-mysql wrapper and add connection retries
the connection is dying with an EPIPE and

    Error: This socket has been ended by the other party

strange that this is only happening with mysql2 and not mysqljs (or
maybe it's a newer version of mysql).  it looks like the likely cause is
mysql's `wait_timeout` system variable.  rather than changing that, just
reconnect if the connection is terminated.